Watched but Unseen
This thesis uses the lens of the interactive installation "Watched but Unseen" to explore the state of people under the pressure of contemporary society and the complex relationship with individual freedom. This article draws on the growing rates of unhappiness, as well as social pressure and the relationship between art and society, to lead to our project "Watched but Unseen". We use symbolic and interactive elements to allow the audience to interact with it in the process. It further allows people to experience the standpoints of various social roles under the pressure of today's society, prompting people to reflect on individual needs in today's society.